# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Nitip Core REST API server.",
        "title": "Nitip Core API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "email": "support@nitip.id"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8000",
    "basePath": "/api/v1",
    "paths": {
...
        "/notifications": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retrieve paginated notifications for the logged-in user",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "[User] Notifications"
                ],
                "summary": "List notifications",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Limit",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Offset",
                        "name": "offset",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_codecoffy_nitip-core_pkg_response.envelope"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/internal_domain_notification.Notification"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/notifications/read-all": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Update all user's notifications to read status",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "[User] Notifications"
                ],
                "summary": "Mark all notifications as read",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_codecoffy_nitip-core_pkg_response.envelope"
                        }
                    }
                }
            }
        },
        "/notifications/unread-count": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get the number of unread notifications for the user",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "[User] Notifications"
                ],
                "summary": "Get unread notification count",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_codecoffy_nitip-core_pkg_response.envelope"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "object",
                                            "additionalProperties": {
                                                "type": "integer"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/notifications/{id}/read": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Update a specific notification status to read",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "[User] Notifications"
                ],
                "summary": "Mark notification as read",
                "parameters": [
                    {
                        "type": "string",
                        "format": "uuid",
                        "description": "Notification UUID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_codecoffy_nitip-core_pkg_response.envelope"
                        }
                    }
                }
            }
        },
        "/orders": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Create an order as a requester",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "[User] Order"
                ],
                "summary": "Create a new order",
                "parameters": [
                    {
                        "description": "Order payload",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_domain_order.CreateOrderRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_codecoffy_nitip-core_pkg_response.envelope"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/internal_domain_order.Order"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_codecoffy_nitip-core_pkg_response.envelope"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_codecoffy_nitip-core_pkg_response.envelope"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/github_com_codecoffy_nitip-core_pkg_response.envelope"
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_codecoffy_nitip-core_pkg_response.envelope"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "errors": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/github_com_codecoffy_nitip-core_pkg_response.ValidationError"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
...
        "/orders/{id}/review": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "View the review left on a specific order",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "[Shared] Order View"
                ],
                "summary": "Get review by order",
                "parameters": [
                    {
                        "type": "string",
                        "format": "uuid",
                        "description": "Order UUID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_codecoffy_nitip-core_pkg_response.envelope"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/internal_domain_review.Review"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/github_com_codecoffy_nitip-core_pkg_response.envelope"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Allows the Requester to rate the Runner after order is completed",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "[User] Order"
                ],
                "summary": "Submit a review",
                "parameters": [
                    {
                        "type": "string",
                        "format": "uuid",
                        "description": "Order UUID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Review payload",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_domain_review.SubmitReviewRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_codecoffy_nitip-core_pkg_response.envelope"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_codecoffy_nitip-core_pkg_response.envelope"
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_codecoffy_nitip-core_pkg_response.envelope"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "errors": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/github_com_codecoffy_nitip-core_pkg_response.ValidationError"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "github_com_codecoffy_nitip-core_pkg_response.Meta": {
            "type": "object",
            "properties": {
                "limit": {
                    "type": "integer"
                },
                "page": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                },
                "total_pages": {
                    "type": "integer"
                }
            }
        },
        "github_com_codecoffy_nitip-core_pkg_response.ValidationError": {
            "type": "object",
            "properties": {
                "field": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "github_com_codecoffy_nitip-core_pkg_response.envelope": {
            "type": "object",
            "properties": {
                "data": {},
                "errors": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_codecoffy_nitip-core_pkg_response.ValidationError"
                    }
                },
                "message": {
                    "type": "string"
                },
                "meta": {
                    "$ref": "#/definitions/github_com_codecoffy_nitip-core_pkg_response.Meta"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
...
        "internal_domain_notification.Notification": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "is_read": {
                    "type": "boolean"
                },
                "message": {
                    "type": "string"
                },
                "metadata": {
                    "type": "object",
                    "additionalProperties": true
                },
                "title": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
...
        "internal_domain_order.CreateOrderRequest": {
            "type": "object",
            "required": [
                "delivery_lat",
                "delivery_lng",
                "estimated_cost",
                "item_details",
                "payment_method",
                "pickup_lat",
                "pickup_lng",
                "service_category",
                "volume_liters",
                "weight_kg"
            ],
            "properties": {
                "delivery_address": {
                    "type": "string"
                },
                "delivery_lat": {
                    "type": "number"
                },
                "delivery_lng": {
                    "type": "number"
                },
                "delivery_name": {
                    "type": "string"
                },
                "estimated_cost": {
                    "type": "number",
                    "minimum": 0
                },
                "item_details": {
                    "type": "string"
                },
                "payment_method": {
                    "type": "string",
                    "enum": [
                        "escrow",
                        "cod"
                    ]
                },
                "pickup_address": {
                    "type": "string"
                },
                "pickup_lat": {
                    "type": "number"
                },
                "pickup_lng": {
                    "type": "number"
                },
                "pickup_name": {
                    "type": "string"
                },
                "receiver_name": {
                    "type": "string"
                },
                "receiver_phone": {
                    "type": "string"
                },
                "service_category": {
                    "description": "Nitip Kirim Fields",
                    "type": "string",
                    "enum": [
                        "beli",
                        "kirim"
                    ]
                },
                "volume_liters": {
                    "description": "Frontend maps S/M/L to liters",
                    "type": "number",
                    "minimum": 0
                },
                "weight_kg": {
                    "type": "number",
                    "minimum": 0
                }
            }
        },
...
        "internal_domain_order.Order": {
            "type": "object",
            "properties": {
                "adjusted_cost": {
                    "description": "Price Adjustment",
                    "type": "number"
                },
                "adjustment_reason": {
                    "type": "string"
                },
                "adjustment_status": {
                    "type": "string"
                },
                "checking_fee": {
                    "type": "number"
                },
                "cod_handling_fee": {
                    "type": "number"
                },
                "completion_code": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "delivery_address": {
                    "type": "string"
                },
                "delivery_fee": {
                    "type": "number"
                },
                "delivery_image_url": {
                    "type": "string"
                },
                "delivery_lat": {
                    "type": "number"
                },
                "delivery_lng": {
                    "type": "number"
                },
                "delivery_name": {
                    "type": "string"
                },
                "dispute_proof_url": {
                    "type": "string"
                },
                "dispute_reason": {
                    "type": "string"
                },
                "disputed_at": {
                    "type": "string"
                },
                "distance_km": {
                    "type": "number"
                },
                "estimated_cost": {
                    "type": "number"
                },
                "id": {
                    "type": "string"
                },
                "item_details": {
                    "type": "string"
                },
                "order_type": {
                    "type": "string"
                },
                "payment_method": {
                    "type": "string"
                },
                "payment_status": {
                    "type": "string"
                },
                "pickup_address": {
                    "type": "string"
                },
                "pickup_lat": {
                    "type": "number"
                },
                "pickup_lng": {
                    "type": "number"
                },
                "pickup_name": {
                    "description": "Recent Pickup \u0026 Proximity Matching",
                    "type": "string"
                },
                "receipt_image_url": {
                    "description": "Phase 2: Proof \u0026 Dispute",
                    "type": "string"
                },
                "receiver_name": {
                    "type": "string"
                },
                "receiver_phone": {
                    "type": "string"
                },
                "requester_id": {
                    "type": "string"
                },
                "runner_id": {
                    "type": "string"
                },
                "service_category": {
                    "description": "Nitip Kirim (Package Delivery)",
                    "type": "string"
                },
                "service_fee": {
                    "type": "number"
                },
                "status": {
                    "type": "string"
                },
                "total_payment": {
                    "type": "number"
                },
                "trip_id": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "volume_liters": {
                    "type": "number"
                },
                "weight_kg": {
                    "type": "number"
                }
            }
        },
...
        "internal_domain_review.Review": {
            "type": "object",
            "properties": {
                "comment": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "order_id": {
                    "type": "string"
                },
                "rating": {
                    "type": "integer",
                    "maximum": 5,
                    "minimum": 1
                },
                "reviewee_id": {
                    "type": "string"
                },
                "reviewer_id": {
                    "type": "string"
                }
            }
        },
        "internal_domain_review.SubmitReviewRequest": {
            "type": "object",
            "required": [
                "rating"
            ],
            "properties": {
                "comment": {
                    "type": "string",
                    "maxLength": 500
                },
                "rating": {
                    "type": "integer",
                    "maximum": 5,
                    "minimum": 1
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    },
...
}